Initial framework for page navigation
