+++ app.json
@@ -3,6 +3,7 @@
     "name": "BrainViewer",
     "slug": "BrainViewer",
     "version": "1.0.0",
+    "scheme": "com.driusan.brainviewer",
     "orientation": "default",
     "icon": "./assets/icon.png",
     "userInterfaceStyle": "light",
