+++ app/viewer.js
@@ -1,0 +1,5 @@
+import ViewerContainer from "../components/ViewerContainer";
+
+export default function Viewer() {
+  return <ViewerContainer/>;
+}
