+++ babel.config.js
@@ -2,5 +2,8 @@ module.exports = function(api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
+    plugins: [
+      require.resolve("expo-router/babel"),
+    ]
   };
 };
