+++ components/SegmentSlider.js
